moove component to other folder
